Create github-actions-demo.yml
This workflow is triggered on every push to the main branch, but not when there are changes to files in the .github folder.
Also, this workflow is triggered manually on every pull request with main as the base branch every Sunday at 6:15 UTC
The build job outputs the name of the event that triggered the workflow and also outputs the name of the branch that the repository is currently referencing. Lastly, it lists all files in the repository.